move address network inconsistent error into address file
