+++ crates/client/src/rpc/types/call_request.rs
@@ -6,8 +6,10 @@ use crate::rpc::{
     error_codes::invalid_params,
     types::{
         address::RpcAddress,
-        cfx::{to_primitive_access_list, CfxAccessList},
-        errors::{check_rpc_address_network, RcpAddressNetworkInconsistent},
+        cfx::{
+            check_rpc_address_network, to_primitive_access_list, CfxAccessList,
+            RcpAddressNetworkInconsistent,
+        },
         Bytes,
     },
     RpcResult,
