+++ crates/client/src/rpc/types/cfx/mod.rs
@@ -2,4 +2,7 @@ mod access_list;
 pub mod address;
 
 pub use access_list::*;
-pub use address::RpcAddress;
+pub use address::{
+    check_rpc_address_network, RcpAddressNetworkInconsistent, RpcAddress,
+    UnexpectedRpcAddressNetwork,
+};
